Entities are sorted to ensure they're drawn in the correct order. Added a second player to test.

+++ Avocado/Avocado/Entity/Enemy.cs
@@ -10,7 +10,7 @@ namespace Avocado
 	{
 		int health;
 
-		public Enemy(Texture2D texture, Vector2 position, int health, int speed) : 
+		public Enemy(Texture2D texture, Vector2 position, int health, float speed) : 
 			base(texture, position, speed)
 		{
 			this.health = health;
